ZEN-30094: removed Discovered device class from defalt list
On a watch list portlet user is not able to remove Discovered
device class.
Occording to new changes we have a regex that add /cse to
all the string in js that has /zport/dmd so remove logic
was not able to match uid in a record and in the grid.
Removed the Discovered from default portlet state since there
is no point in keeping it.